Move `Trace` to the `effect` subpackage
